+++ src/csv/deser/niseci/mod.rs
@@ -15,7 +15,9 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-use crate::csv::deser::{check_path_is_file_ends_with_csv, process_csv_errors, NormalizerReader};
+use crate::csv::deser::{
+    check_path_is_file_ends_with_csv, process_csv_errors, CsvConfig, NormalizerReader,
+};
 use crate::deser::{
     parse_serialized_records, validate_serialized_records, RecordAnagraficaNISECI,
     RecordCampionamentoNISECI, RecordRiferimentoNISECI, TipoRecord,
@@ -133,6 +135,9 @@ where
     parse_serialized_records(iter)
 }
 
+#[deprecated(
+    note = "v0.2 will change signature to expect an explicit delimiter argument.\nConsider using crate::csv::deser::niseci::check_riferimento_niseci_reader_conf() instead"
+)]
 pub fn check_riferimento_niseci_reader<R: Read, T>(
     reader: R,
     has_headers: bool,
@@ -140,8 +145,6 @@ pub fn check_riferimento_niseci_reader<R: Read, T>(
 where
     T: RecordRiferimentoNISECI + 'static,
 {
-    let normalizing_reader = NormalizerReader::new(reader);
-
     let type_id = TypeId::of::<T>(); // Get the TypeId of T at runtime
 
     // Match on the TypeId to determine the actual type of T
@@ -150,9 +153,24 @@ where
         _ => b',',
     };
 
+    let config = CsvConfig::default()
+        .with_delimiter(delimiter)
+        .with_headers(has_headers);
+    check_riferimento_niseci_reader_conf::<R, T>(reader, config)
+}
+
+pub fn check_riferimento_niseci_reader_conf<R: Read, T>(
+    reader: R,
+    config: CsvConfig,
+) -> Result<Vec<T>, Vec<csv::Error>>
+where
+    T: RecordRiferimentoNISECI + 'static,
+{
+    let normalizing_reader = NormalizerReader::new(reader);
+
     let mut rdr = csv::ReaderBuilder::new()
-        .delimiter(delimiter)
-        .has_headers(has_headers)
+        .delimiter(config.delimiter())
+        .has_headers(config.has_headers())
         .from_reader(normalizing_reader);
     let iter = rdr.deserialize();
     validate_serialized_records(iter, |errors| {
@@ -170,9 +188,35 @@ where
     })
 }
 
+#[deprecated(
+    note = "v0.2 will change signature to expect an explicit delimiter argument.\nConsider using crate::csv::deser::niseci::check_riferimento_niseci_path_conf() instead"
+)]
 pub fn check_riferimento_niseci_path<T>(
     path: PathBuf,
     has_headers: bool,
+) -> Result<Vec<T>, Vec<csv::Error>>
+where
+    T: RecordRiferimentoNISECI + 'static,
+{
+    let type_id = TypeId::of::<T>(); // Get the TypeId of T at runtime
+
+    // Match on the TypeId to determine the actual type of T
+    let delimiter = match type_id {
+        id if id == TypeId::of::<VeryItalianRecordRiferimentoNISECI>() => b';',
+        _ => b',',
+    };
+
+    check_riferimento_niseci_path_conf::<T>(
+        path,
+        CsvConfig::default()
+            .with_delimiter(delimiter)
+            .with_headers(has_headers),
+    )
+}
+
+pub fn check_riferimento_niseci_path_conf<T>(
+    path: PathBuf,
+    config: CsvConfig,
 ) -> Result<Vec<T>, Vec<csv::Error>>
 where
     T: RecordRiferimentoNISECI + 'static,
@@ -186,7 +230,7 @@ where
         return Err(err_vec);
     }
     let file = File::open(path).expect("Unable to open file");
-    check_riferimento_niseci_reader(file, has_headers)
+    check_riferimento_niseci_reader_conf(file, config)
 }
 
 #[deprecated(
@@ -249,6 +293,9 @@ where
     parse_serialized_records(iter)
 }
 
+#[deprecated(
+    note = "v0.2 will change signature to expect an explicit delimiter argument.\nConsider using crate::csv::deser::niseci::check_campionamento_niseci_reader_conf() instead"
+)]
 pub fn check_campionamento_niseci_reader<R: Read, T>(
     reader: R,
     has_headers: bool,
@@ -256,8 +303,6 @@ pub fn check_campionamento_niseci_reader<R: Read, T>(
 where
     T: RecordCampionamentoNISECI + 'static,
 {
-    let normalizing_reader = NormalizerReader::new(reader);
-
     let type_id = TypeId::of::<T>(); // Get the TypeId of T at runtime
 
     // Match on the TypeId to determine the actual type of T
@@ -266,9 +311,26 @@ where
         _ => b',',
     };
 
+    check_campionamento_niseci_reader_conf::<R, T>(
+        reader,
+        CsvConfig::default()
+            .with_delimiter(delimiter)
+            .with_headers(has_headers),
+    )
+}
+
+pub fn check_campionamento_niseci_reader_conf<R: Read, T>(
+    reader: R,
+    config: CsvConfig,
+) -> Result<Vec<T>, Vec<csv::Error>>
+where
+    T: RecordCampionamentoNISECI + 'static,
+{
+    let normalizing_reader = NormalizerReader::new(reader);
+
     let mut rdr = csv::ReaderBuilder::new()
-        .delimiter(delimiter)
-        .has_headers(has_headers)
+        .delimiter(config.delimiter())
+        .has_headers(config.has_headers())
         .from_reader(normalizing_reader);
     let iter = rdr.deserialize();
     validate_serialized_records(iter, |errors| {
@@ -286,9 +348,35 @@ where
     })
 }
 
+#[deprecated(
+    note = "v0.2 will change signature to expect an explicit delimiter argument.\nConsider using crate::csv::deser::niseci::check_campionamento_niseci_path_conf() instead"
+)]
 pub fn check_campionamento_niseci_path<T>(
     path: PathBuf,
     has_headers: bool,
+) -> Result<Vec<T>, Vec<csv::Error>>
+where
+    T: RecordCampionamentoNISECI + 'static,
+{
+    let type_id = TypeId::of::<T>(); // Get the TypeId of T at runtime
+
+    // Match on the TypeId to determine the actual type of T
+    let delimiter = match type_id {
+        id if id == TypeId::of::<VeryItalianRecordCampionamentoNISECI>() => b';',
+        _ => b',',
+    };
+
+    check_campionamento_niseci_path_conf::<T>(
+        path,
+        CsvConfig::default()
+            .with_delimiter(delimiter)
+            .with_headers(has_headers),
+    )
+}
+
+pub fn check_campionamento_niseci_path_conf<T>(
+    path: PathBuf,
+    config: CsvConfig,
 ) -> Result<Vec<T>, Vec<csv::Error>>
 where
     T: RecordCampionamentoNISECI + 'static,
@@ -302,7 +390,7 @@ where
         return Err(err_vec);
     }
     let file = File::open(path).expect("Unable to open file");
-    check_campionamento_niseci_reader(file, has_headers)
+    check_campionamento_niseci_reader_conf(file, config)
 }
 
 #[deprecated(
@@ -408,6 +496,9 @@ where
     parse_serialized_records(iter)
 }
 
+#[deprecated(
+    note = "v0.2 will change signature to expect an explicit delimiter argument.\nConsider using crate::csv::deser::niseci::check_anagrafica_niseci_reader_conf() instead"
+)]
 pub fn check_anagrafica_niseci_reader<R: Read, T>(
     reader: R,
     has_headers: bool,
@@ -415,8 +506,6 @@ pub fn check_anagrafica_niseci_reader<R: Read, T>(
 where
     T: RecordAnagraficaNISECI + 'static,
 {
-    let normalizing_reader = NormalizerReader::new(reader);
-
     let type_id = TypeId::of::<T>(); // Get the TypeId of T at runtime
 
     // Match on the TypeId to determine the actual type of T
@@ -425,9 +514,26 @@ where
         _ => b',',
     };
 
+    check_anagrafica_niseci_reader_conf::<R, T>(
+        reader,
+        CsvConfig::default()
+            .with_delimiter(delimiter)
+            .with_headers(has_headers),
+    )
+}
+
+pub fn check_anagrafica_niseci_reader_conf<R: Read, T>(
+    reader: R,
+    config: CsvConfig,
+) -> Result<Vec<T>, Vec<csv::Error>>
+where
+    T: RecordAnagraficaNISECI + 'static,
+{
+    let normalizing_reader = NormalizerReader::new(reader);
+
     let mut rdr = csv::ReaderBuilder::new()
-        .delimiter(delimiter)
-        .has_headers(has_headers)
+        .delimiter(config.delimiter())
+        .has_headers(config.has_headers())
         .from_reader(normalizing_reader);
     let iter = rdr.deserialize();
     validate_serialized_records(iter, |errors| {
@@ -445,9 +551,34 @@ where
     })
 }
 
+#[deprecated(
+    note = "v0.2 will change signature to expect an explicit delimiter argument.\nConsider using crate::csv::deser::niseci::check_anagrafica_niseci_path_conf() instead"
+)]
 pub fn check_anagrafica_niseci_path<T>(
     path: PathBuf,
     has_headers: bool,
+) -> Result<Vec<T>, Vec<csv::Error>>
+where
+    T: RecordAnagraficaNISECI + 'static,
+{
+    let type_id = TypeId::of::<T>(); // Get the TypeId of T at runtime
+
+    // Match on the TypeId to determine the actual type of T
+    let delimiter = match type_id {
+        id if id == TypeId::of::<VeryItalianRecordAnagraficaNISECI>() => b';',
+        _ => b',',
+    };
+    check_anagrafica_niseci_path_conf::<T>(
+        path,
+        CsvConfig::default()
+            .with_delimiter(delimiter)
+            .with_headers(has_headers),
+    )
+}
+
+pub fn check_anagrafica_niseci_path_conf<T>(
+    path: PathBuf,
+    config: CsvConfig,
 ) -> Result<Vec<T>, Vec<csv::Error>>
 where
     T: RecordAnagraficaNISECI + 'static,
@@ -461,5 +592,5 @@ where
         return Err(err_vec);
     }
     let file = File::open(path).expect("Unable to open file");
-    check_anagrafica_niseci_reader(file, has_headers)
+    check_anagrafica_niseci_reader_conf(file, config)
 }